research-agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "pydantic",
    "python-dotenv",
    "pydub",
    "lmnr"
]
